リデューサとコンテキストを合わせて使用
